Deduplicate hook usage and avoid invariant violation
