+++ x-pack/legacy/plugins/infra/public/containers/logs/log_analysis/index.ts
@@ -5,6 +5,7 @@
  */
 
 export * from './log_analysis_capabilities';
+export * from './log_analysis_cleanup';
 export * from './log_analysis_jobs';
 export * from './log_analysis_results';
 export * from './log_analysis_results_url_state';
