+++ x-pack/legacy/plugins/infra/public/pages/logs/analysis/page.tsx
@@ -4,68 +4,17 @@
  * you may not use this file except in compliance with the Elastic License.
  */
 
-import React, { useContext, useEffect } from 'react';
-import chrome from 'ui/chrome';
-import { i18n } from '@kbn/i18n';
+import React from 'react';
+
 import { ColumnarPage } from '../../../components/page';
-import { LoadingPage } from '../../../components/loading_page';
+import { AnalysisPageContent } from './page_content';
 import { AnalysisPageProviders } from './page_providers';
-import { AnalysisResultsContent } from './page_results_content';
-import { AnalysisSetupContent } from './page_setup_content';
-import { AnalysisSetupUnavailable } from './page_setup_unavailable';
-import { useLogAnalysisJobs } from '../../../containers/logs/log_analysis/log_analysis_jobs';
-import { useLogAnalysisCleanup } from '../../../containers/logs/log_analysis/log_analysis_cleanup';
-import { useLogAnalysisCapabilities } from '../../../containers/logs/log_analysis';
-import { Source } from '../../../containers/source';
 
 export const AnalysisPage = () => {
-  const { sourceId, source } = useContext(Source.Context);
-  const spaceId = chrome.getInjected('activeSpace').space.id;
-  const {
-    isSetupRequired,
-    isLoadingSetupStatus,
-    setupMlModule,
-    isSettingUpMlModule,
-    didSetupFail,
-    hasCompletedSetup,
-  } = useLogAnalysisJobs({
-    indexPattern: source ? source.configuration.logAlias : '',
-    sourceId,
-    spaceId,
-    timeField: source ? source.configuration.fields.timestamp : '',
-  });
-  const { hasLogAnalysisCapabilites } = useLogAnalysisCapabilities();
-  if (!hasLogAnalysisCapabilites) {
-    return <AnalysisSetupUnavailable />;
-  }
-
-  const { cleanupMLResources, isCleaningUp } = useLogAnalysisCleanup({ sourceId, spaceId });
-  useEffect(() => {
-    if (didSetupFail) {
-      cleanupMLResources();
-    }
-  }, [didSetupFail, cleanupMLResources]);
-
   return (
     <AnalysisPageProviders>
       <ColumnarPage data-test-subj="infraLogsAnalysisPage">
-        {isLoadingSetupStatus ? (
-          <LoadingPage
-            message={i18n.translate('xpack.infra.logs.analysisPage.loadingMessage', {
-              defaultMessage: 'Checking status of analysis jobs...',
-            })}
-          />
-        ) : isSetupRequired ? (
-          <AnalysisSetupContent
-            didSetupFail={didSetupFail}
-            isSettingUp={isSettingUpMlModule}
-            setupMlModule={setupMlModule}
-            isCleaningUpAFailedSetup={isCleaningUp}
-            indexPattern={source ? source.configuration.logAlias : ''}
-          />
-        ) : (
-          <AnalysisResultsContent sourceId={sourceId} isFirstUse={hasCompletedSetup} />
-        )}
+        <AnalysisPageContent />
       </ColumnarPage>
     </AnalysisPageProviders>
   );
