+++ x-pack/legacy/plugins/infra/public/pages/logs/analysis/page_unavailable_content.tsx
@@ -10,7 +10,7 @@ import { FormattedMessage } from '@kbn/i18n/react';
 import { EuiEmptyPrompt, EuiLink, EuiButton } from '@elastic/eui';
 import euiStyled from '../../../../../../common/eui_styled_components';
 
-export const AnalysisSetupUnavailable: React.FunctionComponent<{}> = () => (
+export const AnalysisUnavailableContent: React.FunctionComponent<{}> = () => (
   <EmptyPrompt
     title={
       <h2>
